pyproject.toml
[tool.poetry]
name = "fastapiproject"
version = "0.1.0"
description = ""
authors = ["Keinz-C <junhe0689@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
gunicorn = "^23.0.0"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.0"
tortoise-orm = {extras = ["asyncmy"], version = "^0.21.7"}
aerich = "^0.7.2"
cryptography = "^43.0.3"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
coverage = "^7.6.4"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"

[tool.mypy]
python_version = 3.12
strict = true
plugins = [
    "pydantic.mypy"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
